add missing tests

+++ commands/remove.js
@@ -29,7 +29,7 @@ export default class Remove extends Command {
     let userId = regExp.exec(text);
 
     if(!userId) {
-      return this.currentPromise.resolve(this._buildPayload("Falto el nombre o pusiste cualquier cosa, no me hagas perder el tiempo. (Agrega @)"));
+      return this.currentPromise.resolve(this._buildPayload("Falto el nombre o pusiste cualquier cosa, no me hagas perder el tiempo."));
     }
 
     return userId[1];
@@ -45,16 +45,20 @@ export default class Remove extends Command {
   }
 
   _remove(userId) {
-    let query = `DELETE FROM players WHERE user_id = "${userId}";`;
+    if(!userId) {
+      return this.currentPromise.resolve(this._buildPayload("Falto el nombre o pusiste cualquier cosa, no me hagas perder el tiempo."));
+    }
+
+    let query = `DELETE FROM players WHERE user_id = '${userId}';`;
     pgConnection.query( query, () => {
-      return this.currentPromise.resolve(this._buildPayload("Todo Pasa."));
+      return this.currentPromise.resolve(this._buildPayload("Todo Pasa..."));
     });
   }
 
   run() {
     let deferred = Q.defer();
 
-    let regExp = new RegExp(/.*(baja <@.*>).*/, "i");
+    let regExp = new RegExp(/.*(baja .*).*/, "i");
     let anotherPlayer = regExp.test(this.payload.text);
 
     this.currentPromise = deferred;
